treeview: introduce watched_iters, refactor has_children_cb
Because we use iters in callbacks, we need to take care of the case where they
get invalidated while the task/whatever is running (between creating & calling
the callback).
Our iters are from the store, so persistent, but the row could have been removed
meanwhile, hence the use of a list of "watched iters" to handle this case.

Also refactored has_children to deal better with different situations, and not
expect that the row has to be EXPAND_UNKNOWN (again, it could have change while
the task was running, by e.g. another task get_children being faster or
something)