treeview: has_children_cb: make sure we're removing the fake node
While under those expand states there should always & only be a fake node, it
could happen (for a short time) that that's not the case. E.g. when
auto-expanding to select the new current location, we might add a node (which
could just replace the fake node) before the get_children() task is being
started; and until it's done the expand_state will not be updated.