Remove DonnaSharedString
After more reading & testing, turns out this whoel idea of shared string wasn't
really a good one. Sure, there is less memory needed, but it shouldn't be that
much in the end (we rarely need that many strings at the same time in multiple
threads/places), it's more code to write/maintain, and if there's a performance
impact it isn't a good one: atomtic operations are slower.

See http://jjacky.com/2013-03-16-strdup-free-vs-atomic-reference-counting/
for some test results.