Add a columntype-data to be used for render & node_cmp
When a column is created in treeview, it will get_data from the columntype. This
pointer will then be used (provided by the treeview) on render & node_cmp, thus
allowing to perform the operation much faster (than having to get the config,
etc)
Since render and esp. node_cmp are very frequent operations, this is a much
needed improvment.