treeview: Reworked the whole select-arrangement thing
Instead of relying on a function on app, we emit a signal. This is much better
as it puts the whole arrangement selection logic out of treeview, where it
doesn't belong.

We also now support arrangement columns, and/or sort, and/or second sort. This
makes things much better for an arrangement that only changes/defines the sort
order. (We also fixed some things re: sort orders, e.g. setting a default one,
(trying to) preserve the current one even when changing columns, etc. And called
everything second_sort_* instead of sometimes secondary_sort, or
sticky_sec_sort, or whatever mess it was)