treeview: if select-arrangement if the same, it's not always applied
We remember the current arrangement, and only apply the one from
select-arrangement if it's a different one, or when the flag ALWAYS was set.
This allows to e.g. apply a "default" sort order, but not reset it back
everytime we change location (to one with the same arrangement, e.g. in case of
arrangement for fs:/*)

Also select-arrangement is only emitted in mode list, since it doesn't apply for
tree. Tree simply load the default arrangement, i.e. under
"treeviews/<NAME>/arrangement" or "defaults/arrangements/tree"