treeview: Fix some loading arrangement loading
After a signal select-arrangement, we can have an "incomplete" one (or none at
all). We then load defaults for all missing elements, making sure we do have
something (except maybe second_sort, always optional).

This will ensure that on first time, we have columns & sort. It also makes sure
than we don't just keep whatever was last used, but fallback on defaults when
nothing was specified on the arrangement, as would be expected.