Add (filters &) color filters
Color filters are added as part of the arrangements. This allow for per-location
color filters, using the type property we can even enable/combine/disable at
will, etc

Color filters are loaded from new config_arr_*() functions, as are all other
elements of the arrangement. Similar to the config_*_column() functions, this
makes things easier and consistent.

Color filters are just wrapped around filters, a more common tool that we'll be
able to reuse as is for other things: visual filters, selection filters, and
even our own search feature!

Filter handles the boolean logic, but each "block" of criteria is linked to a
column, and the columntype handles the matching part.

Filters are loaded from app so they can be used, since we could have the same
filters in multiple color filters, or in color & visual filters, etc (this saves
a bit of memory, plus the time of parsing the filter string)