treeview now handles clicks on its own
That is, we decide whether a click is a double-click or not, setting up timeout
& whatnot, instead of relying on GDK_2BUTTON_PRESS or letting GtkTreeView
trigger the row-activated signal.

This means we have full control over the whole thing, allowing:
- dbl-click being processed w/out a single-click before (i.e. yes, there's a
  little delay after the click before it is processed, waiting for the dbl-click
  timeout to expire). Pretty cool: dbl-click on tree to collapse/expand w/out
  changing location.
- we support slow-double click: a second click after the dbl-click delay and
  before another one. This time the single-click will be processed obviously,
  but this can still be useful to go into edit mode (e.g. rename files).
- we support single, double & slow-double for left, middle & right clicks.

Single-left clicks on main column are handled by the treeview, which also deals
with Ctrl/Shift-clicks on expanders. About everything else gets sent to the
columntype (via handle_click) where it can be processed or not. If not, we might
have some default (e.g. w/out full_row_select, a single-left click will
unselect).
Columntypes can use the given pointers to put the renderer into edit mode, for
in-line editing and whatnot.