treeview: list: Ignore click that are getting the focus
List only, because we expect the tree to not have the focus yet have its single-click action work. List on the other hand it's best to preserve selection and just grab the focus.

Note that a click on tree that gives the focus to the *app* (i.e. app's main window) will still be ignored, however.