provider: add trigger_node_task; Implemented in provider-fs
trigger_node() only works on ITEM, because CONTAINER are just that, container.

The idea will be we'll have a command trigger_node that takes a treeview as well
a a node as parameters, and calls the provider's trigger_node() only on ITEM,
for CONTAINER it just sets it as new location for the treeview.