treestore: Add some extensions to GtkTreeModel
Add ways to get previous/next element, regardless of level/depth. That is, the
node that would get focused on treeview if user pressed up/down. Similarly adds
a way to get the last node, e.g. pressing End on treeview.
Also adds count() to return the number of nodes in the model.

Note that all those are "extensions" to GtkTreeModel, i.e. apply to/account for
visible nodes only.

Also add a version of count() for all nodes in store.