treeview: Add TreeRow, TreeRowId, selection(), set_focus() & set_cursor()
TreeRow will be a way to identify a specific row with certainty even in tree,
where multiple rows can exists for the same node. Will be used e.g. in commands
to id the clicked one.

TreeRowId is a general way to identify a row using either:
- a TreeRow, as described above.
- a Node (first match in case of multiple one in tree)
- a path; that's a regular GtkTreePath (e.g: "0:2") and a few special ones of
  our own:
  :all = all rows
  :selected = selected rows
  :focused = focused row
  :prev/:next = previous/next row, as when using up/down on view
  :last = last row, as when using End on view
  :prev-same-depth/:next-same-depth = previous/next row on same level

set_focus() doesn't exist with vanilla GTK, because it's too much of a PITA to implement.