Add command support & provider-command
All the work is done in command.c so that it can be used from outside the
provider. This is because e.g. treeview will do some parsing (well, not parsing
but some converting of args) when processing its clicks, to bypass a few steps
(such as running a task to get a node for the command, etc).