treeview: Refactor click handling, got full control, use options to trigger nodes
Now all the clicks done in a treeview are simply matched to an option, e.g.
left_click or ctrl_middle_double_click, which sets which node to be triggered in
responde to the click.

Those options cover the 36 kinds of clicks supported (left/middle/right;
std/double/slow; and with Ctrl and/or Shift held) are can be defined in for each
column (even in arrangements!), or the treeview for some "local defaults".

Also, for each option, in the treeview, we can have them prefixed with "blank_",
"blankrow_" or "blankcol_" when the click was in blank space (i.e. not a
renderer), below the last row, or in the blank column on the right (mode list
only). Should be noted that (only) the blank_* options can be set per columns.