treeview: tree: Add "clicks" to define per-item actions on clicks
In mode tree, mostly useful for bookmarks and such, we had a way to define
actions on clicks specific to one item. It is basically like a tree_visual,
expect it's not one (internally it's handled as one).

This could allow to e.g. have bookmarks of ITEM that on click are triggered,
expect one that gets set as location in the list, etc.