treeview: Fix collapsing parent of current location leaving no selection
This was bad because we require a selection/expect in such a case to have a
change of location (in sync_with) to occur. Also could lead to a segfault as our
location_iter wasn't invalided, yet could now point to a non-existing row (w/
full collapse).

Now, collapsing the parent of current location sets it to the focused row. Note
that collapsing parent of focused row will make that parent the new current
location, as per GTK behavior where since the cursor is gone, a set_cursor() is
triggered. I'd like a simple set_focused_row() better, but.