command: Update arg parser
Now for string argument from a list, instead of an exact match a partial match
is enough. Meaning, to match "foobar" all you need is "f" *if there are no other
items starting with an 'f'*; so with "foosome" and "foomore" as possibilities,
"foo" isn't enough to have a match, but "foos" and "foom" are.

Checks are also case insensitive, so in the last example, one could use e.g.
"fooS" or "fooMo"