config/columntype-value: Add new type of extra: list-flags
Pretty much exactly like list-int, except the acutal (int) values must be >0 and
power of 2, since they're to be flags. As in, the value of an option can be any
combination of the supported values.

In file, values are listed as a comma-separated list. In ct, the labels are
listed similarly, and a window of checkboxes is shown when editing.