treeview: Fix segfault when filters asked for ct_data of column not in tree
This is of course totally legit, as one can filter via any column without
needing said column to be used on tree. So we load a ct_data for the tree (using
current arrangement, etc ofc) of use that This stays in memory until next
load_arrangement, so we can keep using the filter with that ct_data.