donna: Add missing locks to access filters, visuals, columntypes & providers
Those can actually be accessed (and, therefore, loaded) from outside the main
thread, so we need to have locks.

Filters are handled via toggle_ref so we can keep them alive for a little while
(15m) after the last unref, and only then unref them. Nice when re-using the
same filter, which happens a lot when changing location (free-ing arrangement &
selecting a new one (or the same))