Change parsing of locations; Add notion of "internal references"
The idea behind those "internal references" is for use in scripts (i.e. external
processes), where they can get arguments (from command-line parsing or as
command's return value) and use them in other commands. Saves having to reload a
node, or worse when dealing with e.g. an array of nodes.

Dereferencing means %v resolves to such an intref, %*v means using a string that
can also work the same, would work on another session on donna, but needs more
work. E.g. a full location for a node, or a list of full locations for an array
of nodes.

Similarily, when external processes will have donna run a command, they'll be
able to prefix it with * to dereference the returned value (else an intref), or
with # to have it ignored (nothing gets returned).