treeview, list: Handling click checks under "selected" first
In mode list only (selection is quite a different thing in mode tree, and it
wouldn't really make sense there), the clicks will be looked under subcategory
"selected" first (if nothing found, not even defaults, then we fallback to
regular places) if the clicked item was selected.

Allows to have e.g. dbl-click activate the node if not selected, but all
selected nodes if selected. (Or, you can have it always activate the clicked
node, selected or not, and use another click to activate the selection.)