command: Allow _donna_command_parse_run() to be blocking
In order to handle keys with a motion spec, we'll have to trigger the "motion
action" (via _donna_command_parse_run) and then trigger the actual/main action.
But, if that happens in another thread leaving the user free to change
selection, location, etc this will be a huge mess.

So, in that particular case we'll need to have it blocking, and return whether
it was a success or not (from the GUI thread).