treeview: Add our own handling of keys
The way keys are handled in treeview isn't the usual GTK behavior, but uses a
vim-inspired system. That is, pressing a key can trigger an action right away,
or it can wait for a "specifier" (SPEC).
It is also possible to have specified a "multiplier" before hand.

The multiplier is simply a number (one or more digits) typed in before the
actual key. A spec is simply a key, and can be a lowercase letter, an uppercase
letter, and/or a digit. It can also be a "motion".

A motion is an action classified as such, which indicates that it will move the
focused row. Using a motion spec allows keys to work on a range, much like via
clicks: between the focused row & the clicked row.
Here, between the originally focused row, and the one after the motion has been
processed.

Multiplier and spec will simply translate into new variables parsed in the
full location to trigger, %m and %s respectively. This means the multiplier
doesn't automatically work an everything, repeating the action as many times as
speicified, but requires the %m to have been used, and be understood, by the
command.

In addition, an extra spec - called "combine" - can be used. That is, the first
key pressed will be a combine accepting a spec. Then another key must be
pressed, and its full location will have an additional %c resolved to the
combine-spec.

The combine key sets its name (e.g. "register") and only action with a matching
combine option set will be allowed to be used with that combine.

Every key can be of a different type:

- COMBINE: takes a spec to be the combine of the following action/key

- DIRECT: trigger the action right away

- SPEC: needs a spec to trigger the action

- ALIAS: an alias to another key (i.e. both keys behave the same)

- DISABLED: the key will do nothing