command: Add repeat to repeat a command multiple times
Should be noted that this might not always work/be a good idea, depending on which command one wants to repeat, particularly because repeat will be processed in a separate thread.

Also note that since it's argument for command to be repeated in a string, any %flag to treeviews or nodes will go through intrefs.