app: Add signal treeview-loaded; Remove unused active-list-changed
Because on app start treeviews will not necesarilly be loaded "in order" it'll
be possible for a treeview to want to sync with a tree not yet loaded. To handle
such cases properly, they listen to this new signal.

active-list-changed is an old thing from the past, active-list is a property
now.