Add statusbar widget, using status provider-s
One adds "areas" to the statusbar, by specifying some options (e.g. natural
size, whether it expands to additional space or not...), a status provider & id
for that area.

The provider will specify which GtkCellRenderer-s are needed in that area, and
take care of setting properties on the renderers (or setting the tooltip). It
will also emit a signal (status-changed) when a refresh is needed.

It is possible to update an area, meaning set a new provider/id for that area.
The renderers must obviously be the same ones, as the idea is to switch from one
provider to another of the same "kind", e.g. from one treeview to another when
the area must be filled from the active list.