treeview: Fix issues when collapsing/removing focused/current row
Removing a row that had the focus could lead to GTK doing a set_cursor, and in
tree could also lead to an unwanted change of current location. This is now
properly handled, as we check whether the row to remove is focused/current, and
if so change things before the removal.

Similarly, when collapsing a row and the focus/current was below could also lead
to issues/unexpected behaviors. This is now properly handled in
test-collapse-row, which will again move the focus/selection before the
collapsing, thus avoiding any GTK's set_cursor.