treeview: Properly handle focused row being "fake removed"
While we handle the removal of rows, there's one case where a row could be
"removed" from the tree outside of remove_row_from_tree() : when it gets hidden
(through visual filter, show_hidden option, etc)

In this case, we need to handle the case when it has the focus, to properly move
it (before GTK tries to call set_cursor). This couldn't be done via row-deleted
because either GTK had already done its set_cursor call, or the row was gone
from the model but not the view yet, preventing from being able to go prev/next
(since the view still used "old paths").

A signal row-fake-deleted as therefore added to our store, to be emitted just
before actually "removing" the row.