treeview: Add a replacement set_focused_row() when not patched
It much uglier/more work than the patched version, but it should produce the
same results. Because I really like to be able to set the focus only, and it was
getting too annoying to maintain different versions for both cases, and w/
different results since the non-patched would often just set_cursor.