treeview: list: Rewrite/fix the changing of location
Changing location in list is really a multi-step process, and should be treated
as such. Here's how it goes, most of it done in the new change_location() :

- we ask to change location / start a get_children task. We also switch
  providers (i.e. listen to signals of future_location instead of current one)
  because this will allow is to get children (via signal new-child) as the task
  is running. This is mostly useful for search results.

- (optional) the task is "slow" so we clear the store & show the "please wait"
  message on the view

- (optional) we got a child (via new-child), we clear the store (unless we were
  already at SLOW), switch the class domain-*, build the new arrangement,
  officially move (switch location & emit the notify) and add the node.

- task get_children is done: unless GOT_CHILD we need to clear the store, switch
  class, build arrangement. Then we fill the view, but if there was GOT_CHILD we
  need to check for dupes.

This will allow us to do everything we need, including handling search results
as needed (i.e. filling up as the search goes).

Also changes the way a refresh normal is done, where instead of getting a new
get_children task, we try to duplicate the old one (which we kept, if it is
duplicable). This is important for search results, to run the exact same
command, including workdir.

Because usually commands will use the current location (of active-list) as
workdir, but if we're not in "fs" then it just fails.