command: Fix possible segfault or whatnot
When running a command from an action, and the command's task needed to be run
in another thread (than current/UI one, i.e. visibility not FAST nor GUI) then
we would, in the main thread, free rc_data (including the args array for for
command's function!).

We'll now alloc a cr_data to have the array, and free it only when the task has
finished. We don't need to malloc if visibility is GUI or FAST, since then the
task will run in the current thread (which we know to be the UI one).