provider: API change, no more node_remove_task() but a single io_task()
This handles COPY, MOVE and DELETE operations. The sources *must* all be within
the same domain.

This idea is that if the provider supports the operation, it'll return a task to
do it, else NULL. The caller should ask the provider of the source(s) first, and
in case of ERROR_IO_NOT_SUPPORTED try the provider of the destination.

For DELETE obviously there is no destination.