Start adding register support
Registers are named lists of full locations. We only store the full locations, so any change (move/rename/etc) isn't taken into account, as is the norm.

It is possible to create new register, specifying the type (cut/copy). One can also add nodes to an existing register, or change a register's type. A register's type can also be ignored, and a copy or move be performed regardless. One can also removes the nodes from a register.

Because this is a multi-threaded app, when getting nodes from a register one should specify whether the register should be dropped or not, based on what file operation will be done, to avoid issues.

Currently registers aren't loaded/saved, that will come. We'll also add special handling/support for register '*' and '+', which are PRIMARY and CLIPBOARD respectively (a-la vim).