treeview: Add rubber banding support (GTK_IS_JJK only)
Because we don't let GTK do handle the clicks, we can't let it do the rubber
banding either. However, our patchset will allow to manually start a rubber
banding operation, so we can support it.

The way it works is, on a xxx_click option one simplt specifies "rubber-banding"
(instead of a FL), and a rubber banding operation will then be started. (It
obviously should not be triggered w/ xxx_click_on_rls since that will lead to
unexpected results.)

Because of how we do things, there's really only one modifier that has an
effect: CONTROL. Without it, rubber banding selects, adding to the selection.
With it, it inverts instead of selecting.

We've also taken care of making things as good & consistent as possible, even
without delays before clicks are registered. We also don't have the focus moved
during/after a rubber banding operation.