treeview: Update rubber-banding for latest patch
The patch now has GTK not ignore the first row, and deal with special behavior
(toggle if Ctrl) in the rubber-banding-active signal. So we can simply override
this to disable that behavior.

For now that's all, later we'll have an "event" of our own, so special handling
can be done as well.