Rewrite parsing/running commands; Update treeview accordingly
Now the only way to run a command is via its trigger_node task. Parsing happens
as first step in the worker, which then calls the command's function. This is
done directly, or in a new thread if required (i.e. command is GUI but current
task isn't).

The parsing was also redone: now the caller (treeview, menu, etc) will convert
all flags into the FL, using intref if needed. Once that's done, we can get &
trigger the node.