app/donna: get_current_dirname() now returns last path in fs
Because often times one might need an actual "current location" in fs, this will
do the trick. It returns the current location (of active-list) if it's in fs,
else it returns the last one used.

If one needs the "real" current location, get get_current_location() to get that
node.