donna: Fix loading config
Unlike conf-def & CSS, there's no merging for the config. We just load the user
config, and if there's none we (try to) copy the system one over, then load it.

We also keep another copy of the system config as "reference" so later the user
will be able to do a merging of a new system/default config, and have any new
settings imported, as well as those where value in ref & current is the same
(i.e. user didn't change it) but different in new.