provider-base: Add unref_node(), called before a node gets finalized
This new virtual function gets called when removing a node from our hashmap
(thus removing the last ref on the node, which then gets finalized), in case the
provider needs to do some memory cleaning on his own.

Obviously this should be for memory cleaning only, should be as simple/fast as
possible (happens while provider-base is locked) and should never add (or
remove) a ref on node.