Fix providers, handling icons (via gtk_icon_theme) in their own threads
Setting an icon on a node seems like nothing, but those gtk_* functions to
access the icon theme still require to be in the main thread. Not doing so would
sometimes work, other times lead to different kinds of crash.

To do this properly, provider-base now provides a little helper that will take
care of things (since it sets a property a signal is emitted, hence should never
be called from a *locked* provider).

Every provider has been updated to use this (inc. config, via a private call
since it's not a provider-base). fs does things on its own because it doesn't
use a named icon, but gets it from the MIME type.