provider: Add remove_from() to remove nodes from somewhere (not delete them)
Difference being: deleting a node is an IO operation, meaning to "destroy" the
item behind the node (file, option, etc); Whereas removing a node from somewhere
can simply be removing it from a register, virtual folder, etc

Of course, on fs remove a node from its parent is the same thing as deleting it.
In fact, remove_from() tasks only exists in FLAT providers, and when possible
will be "auto-converted" to IO_DELETE task on a non-flat provider.