contextmenu: Fix using unparsed FL
Context-menus are a bit weird, in that usually this parsing in only needed for
actions (keys, clicks), so we parse & then trigger one FL, w/out delay.

Here, we parse a bunch of FLs, put them all in a meny and only one of them (if
any) will be triggered.