treeview: Start adding context menu support
Currenly only has one section, "go"

Extra can be used to specify which item(s) to include, and in what order, as a
;-separated list.

Items are "up" (Go Up), "down" (Go Down) or "-" for a separator. "up" & "down"
are only included on non-flat locations, and "up" is not sensitive on root.