command: Remove node_activate(); Add node_trigger()
Because it became a little more powerful. Now instead of a normal vs alternate
mode, one can specify which action will be done on item & on container.

Item has the same: trigger, and goto.

Container still has goto and popup, but also trigger. In which case, we look for
a (string) property "container_trigger" which must be the FL to trigger. This is
for e.g. internal nodes in context menus an such.
For containers, trigger can be set with either goto or popup, as a fallback if
there's no property container_trigger.