columntype: Add can_edit() to know whether edit can be performed
Doesn't guarantee editing could be succesfully performed, but has to check if ct
supports editing, if the node has the property and if it has writable.

This will be used by treeview to know whether or not to include a (sensitive or
not) item in its context menu. Hence the label arg, used as menu label.