contextmenu: No more sections, but aliases
Now the context_menu option (or the items arg given to the function) is a
coma-separated list of items. An item can be of different form, based on its
first character:

- ':' for internal items, provided by the context source (e.g. treeview)
- '!' for internal aliases
- '@' for user aliases
- else, a user item

An alias must be a string of (self-contained) items (i.e. it cannot open/close a
submenu from outside of it). It can contain submenus, user/internal items as
well as internal aliases. It cannot, however, contain user aliases.