Add task-helpers
Provides an easy way for a task worker to block its thread while some user
feedback is asked for on UI thread. Also handles aborting UI things & unblocking
the thread if it gets paused/cancelled.

A (very) basic way to ask questions/confirmations is provided, in addition to
the base helper.