provider-fs: Add support of "engines" to perform the IO; w/ engine 'basic'
Engines will be what actually performs the IO task (copy/move/delete) in fs.
This will allow to use different tools based on user preferences, as in the user
choose its preferred (default) engine, but also (later) the ability to use
different engines based on the job to be done.

Engine 'basic' is a simple "wrapper" around cp/mv/rm.