task-process: Add class functions for signals; Use them for pipe-new-line
pipe-new-line is now handled from pipe-data-received in the class function,
which will allow it to process extra data when a signal is emitted outside of
data read from the process.

This obviously shouldn' be (ab)used unless there's a very good reason, e.g.
during IO to show the answer (and add the \n so pipe-new-line is emitted).