provider-command: Better handle things when @command returns an array of 1 element
In such a case, if the arg is of matching type (or compatible, i.e. a ROW_ID for
a NODE) then we can put it right away instead of converting to string.

If we do convert to string, put it unquoted since it doesn't hurt arg that
expect/handle arrays, but will make things work (better) for those that don't
(since they expect to deal with unquoted value).