Add provider-invalid and marking node "invalid" if needed
After signal node-deleted has been emitted on a provider, a callback will be
triggered in thread UI. If by that time there are still references on the node,
the provider's unref_node() will be called - to have the node removed from the
provider, including toggle_ref, as if the node didn't exist anymore - and then
it will be marked invalid, i.e. all its properties are removed, the name becomes
"[invalid] <old FL>" and it is now "owned" by provider invalid, which supports
nothing at all.

This is meant for cases where after signal node-deleted, references are still
held on the node. This shouldn't happen, but it could (e.g. as return value of a
(public) task). But the provider shouln't keep that node around, since if a new
file was created with the same name, there could be confusion/problems.