command: Add node_get_property()
For properties of type (u)int(64), an optional formatting option can be used. It
can be either "time" or "size" to indicate the option is a timestamp/size, and
have it formatted using default options.

One can also add "=fmt" to specify a format to use instead of the default, or
use "@category" to specify the config category to get all options from.

Note that when specifying a format, the percent sign needs to be "escaped"
obviously, e.g: date=%%Y