treeview: contextmenu: Add support for alias/items from columns
Items & aliases column.<COLUMN>.<NAME> can be used. The rule is that every
column type should support an alias "options" to show the current options on the
column.

As a convenience, the tree provides an alias "column_options" for the alias
"options" of the current (i.e. from reference) column.