contextmenu: Add option is_container
When set, the node created will be a CONTAINER instead of an ITEM, and
the trigger is expected to be a CONTAINER as well, whose children will
be seen as our node's children.

This simplifies the user TYPE_CONTAINER a lot, since we don't have to
create the node there, setting all possible properties from info, etc

If the item is a context submenu, then we force menu-submenu to ENABLED
since it can't be triggered (we also don't set a container-trigger).