tree: Fix finding closest iter to prefer current branch
In sync, we might be looking for the closest iter to the list's new
location, e.g. with KNOWN_CHILDREN.

The rule was a visible match is gold, then we favor match, visible
non-match and non visible non-match. However, for two results of the
same "scope" we just used the first one found.

We now prefer results from the current branch/root (or the one with the
focus if there's no current root).