tree: sync: Improve getting closest iter
When getting closest iter, e.g. finding the row to focus in NODES or
KNOWN_CHILDREN mode w/out a match, when there's two rows with the same
state (in terms of whether it's a match and/or visible) and none is in
the current root, then we favor the "closest" one to the actual node.

That is, between "/tmp" and "/tmp/foo" the later is closest to
"/tmp/foo/bar"