provider-config: Add note about handlers of new_node()
Specifically, they should never try to write on the config, since that
will deadlock. This is due to new_node() being emitted while the config
is under a reader lock.

Alternative would be to put a non-ready node in the config, then emit
new_node and mark it ready afterwards, which feels/would be worse.