config/command: Add set_option()
This allows to set/create an option, specifying its type and then, for
extras, using the "in-line" values, i.e. same as in the config file.

It also providers a UI to set the type, name & value all at once. This
can use default values given to the command, or imported from the
existing option if any. (Note that it is not possible to change an
option's type, trying to set an option w/ a different type will simply
fail, same as using the set_*() API.)