treeview/command: Add tree_{load,save}_list_file
Allows to load/save a treeview in mode list from/to a file. This
includes current location, focused row, sort order, scroll position
(as a percentage), and selection.

It is possible to define which of those (except required current
location, ofc) will be loaded/saved.

(This will be used for tabs later on.)