treeview: Fix handling removing root of current location
Now we check if the focus & selection (i.e. current location) are
children of the row being removed, and if so handle it right away.

The benefit from letting it handled when the actual children is removed,
is that we can move it once where it shall go, instead of moving it to
the parent, which is then removed so it goes up, and so on until it
reached back the row originally removed.

This also avoids making the list (sync_with) change location when
there's no need for it, which could also lead (via sync) to the tree
re-adding a root.