treeview: Some fix in remove_row_from_tree()
Mainly this is about moving code: we cannot remove the iter from our
hashmap & list of roots before removing the row, because we still need
said iter (e.g. priv->roots might be used in sorting when the model
determines, when removing row, if it has a next sibling).

So this is moved afterwards, where it should be noted that the iter can
still be used with itereq() but not with the model, since the row has
now been removed.