treeview/command: Add tree_{load,save}_tree_file
Allows to load/save a treeview in mode tree from/to a file. Much like
list file, it is possible to define which tree visuals (if any) are to
be loaded/saved.

Tree file contains all roots, all children partially expanded row (in
minitree), which rows were in EXPAND_FULL (i.e. had all children loaded)
and which were expanded, as well as the specified tree visuals (including
for non-loaded rows).