treeview: Add rowid's path "item", "container" & "other"
To refer to the next item, container or opposite of the focused row.

This will start at the focused row, go next until it finds a matching
row. If none found when the bottom is reached, starts back from the top.

This allows e.g. to toggle from items to containers (when sorted apart).