donna: Add new dereferencing mode "fs"
When using a context variable (%X in a trigger/FL of an action), it was
possible to "dereference" it, i.e. for rows & (array of) nodes to use
their full locations instead of internal row format/intrefs, by using
the %*X format.

Now it is possible to use %:X for a "FS dereferencing" which will put
the location when in domain "fs" and nothing for other domain. Makes it
easy to use things in e.g. scripts where we want the actual file names
(i.e. locations) and don't support anything outside of fs.