Fix nodes marked invalid not unref-d by their providers
Since c41467e4 providers wouldn't unref nodes marked invalid, therfore
leading to all sorts of trouble when a new node for the same location
was needed (since they would re-use the invalid node).

This is because since we don't use our own counter in the node, but its
GObject->ref_count, we need not to check its value in this special case.