taskui: Change how functions are meant to be called
Instead of expecting all functions to be called from the main/UI thread,
only show() is. Others are intended to be used by the worker directly
from its thread.

It's up to the taskui to switch to the UI thread (if needed) or use locks
or whatever. Seems to make more sense/be simpler that way.