treeview: load_tree_file: Fix loading expand PARTIAL
When loading a PARTIAL expand state, there was an issue if the first
child added wasn't actually visible. The model/view would then consider
the parent to not have children anymore (since the first child "replaces"
the fake node, and it not being visible "means" the row is removed) which
would result is the PARTIAL expand state lost, as well as collapsing the
row if it was expanded.

We now force keeping the fake node around while adding children, and
remove it once done. This ensures the PARTIAL & expanded state to be
preserved.