treeview: Fix dbl-click not always processed correctly
Especially in case of left clicks, where the single-left occurs right
away, when that resulted in queuing a redraw operation, if it took long
enough the timeout would expire, and both it & the second click would be
in the main loop at the same time.
Since they had the same priority, when the timeout gets processed first
the dbl-click is "ignored" (seen as a slow one).

Giving the timeout a slightly slower priority fixes the issue.