treeview: Fix rubber band glitch on dbl-click
Sometimes on a dbl-click, we would get a quick rubber band operation
and a few rows selected. It was easy to do & all over the place (i.e.
selection would go much broader than the visible rows even).

I don't know why, so this isn't a fix per-se, but by only "starting" the
rubber band if there was no previous event (unless it expired) then we
basically "solve" the issue.