Update debug macros
We now include a(n optional) name for each source. This will be used to
not only specify which source we want debug messages from (e.g. treeview)
but also limit which ones (e.g. only "list").

This obviously doesn't work for APP, TASK, etc but does for TREEVIEW as
well as NODE & PROVIDER (where the domain is used).