donna: Add options --log-level, --verbose & --debug
To specify the minimum log level to show (and because -vvvv is common
and simpler/quicker to type), as well as debug "filters" to specify
which source(s) we want messages from.

Sources are comma-separated, and can include (after a colon) a
plus-separated list of names; E.g:

  --debug task,node:fs will get messages from tasks as well as nodes
from "fs" (ignoring nodes from other domains)

  --debug treeview:foo+bar will get messages from treeviews "foo" and
"bar" only, ignoring any other.