Fix a bunch of warnings (Reduce use of bitfields)
Fix the many warnings that existed, from unused variables to actual
errors. All the while, the use of bitfields was reduced a lot, mostly
because using them was creating complications for not a lot of gain.

That is, to use them without warning we'd need either:
  data->option = (TypeCast) value & MAX_VALUE;
or even (because gboolean is signed) :
  data->is_flag = (guint) value & 1;

Bitfields are still used for internal flags or in node/task as those are
quite common/used a lot. But for treeviews & columntypes it might not be
worth the trouble.